requirements.txt
numpy
pandas
pymilvus
pinecone-client
streamlit
flask
flask-cors
flask_restful
sentence_transformers
gunicorn